+++ examples/drivers/fs-demo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # fs-demo
 
+## 4.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @objectql/core@4.0.1
+  - @objectql/driver-fs@4.0.1
+
 ## 3.0.1
 
 ### Patch Changes